Create basic plugin and override the license link in the dataset form
Basic plugin that follows the theming tutorial in the ckan
documentation. The plugin adds an additional directory 'templates' where
ckan will look for jinja2 templates.

There is one file in this directory which overrides the ckan template of
the same name/directory structure that contains a block override to
replace the open definition link with a creative commons nz link.