Add a 'Terms of Use' link to the front page.
Since we are adding an additional page instead of overriding what is
already there we have to add a controller which just renders our
terms_of_use.html.

The controller has to be mapped to ckan's routing so we update our
plugin so we add an additional route '/terms_of_use' that maps to our
custom controller TermsController. We've named the route so we can refer
to it later when we override the header link.

The terms_of_use.html has the text copied from the original, but we've
extended ckan's default page.html that way the page will fit in with the
rest of ckan and have links back to the dataset page etc.

We've overriden a ckan template 'header.html' and overriden the block
that controls the links at the top and removed the 'About' link and
changed the link to our 'Terms of Use' link by using the name of the
route we defined earlier